Remade the regex that fetches wifi names from netsh wlan show networks
The regular expression fired on whitespaces, causing wifis with white spaces in them to be broken up and the following code could not connect to the network.

i.e. this is my wifi became wifi, since the last white space were the "trigger".
Replaced that with (SSID\s\d{1,3}\s\:\s) to be a much more specific pattern.